humid code revised again

+++ main_read.py
@@ -81,12 +81,12 @@ def collect_humid(bus, temp_comp):
 	#Humid Calculation
 	temp_comp = int(temp_comp)
 	var1 = int(hum_adc) - int(int(par_h1) << 4) - (int(temp_comp * int(par_h3) / 100) >> 1)
-	var2 = int(par_h2) * (((temp_comp * int(par_h4)) / 100) + (((temp_comp * ((temp_comp * int(par_h5)) / 100)) >> 6) / 100 + (1 << 14))) >> 10
+	var2 = int(par_h2) * int(((temp_comp * int(par_h4)) / 100) + ((int(temp_comp * ((temp_comp * int(par_h5)) / 100)) >> 6) / 100 + (1 << 14))) >> 10
 	var3 = var1 * var2
-	var4 = ((int(par_h6) << 7) + (temp_comp * int(par_h7) / 100)) >> 4
-	var5 = ((var3 >> 14) * (var3 >> 14)) >> 10
-	var6 = (var4 * var5) >> 1
-	hum_comp = (var3 + var6) >> 12
-	hum_comp = (((var3 + var6) >> 10) * 1000) >> 12
+	var4 = int((int(par_h6) << 7) + (temp_comp * int(par_h7) / 100)) >> 4
+	var5 = int((var3 >> 14) * (var3 >> 14)) >> 10
+	var6 = int(var4 * var5) >> 1
+	hum_comp = int(var3 + var6) >> 12
+	hum_comp = int((int(var3 + var6) >> 10) * 1000) >> 12
 
 	return hum_comp
